Refactor:
- Extracted all WebSocket Clients and Handlers to websockets package.
