+++ src/main/java/com/suvikollc/resume_rag/websockets/AzureVoiceLiveWebSocketHandler.java
@@ -1,4 +1,4 @@
-package com.suvikollc.resume_rag.serviceImpl;
+package com.suvikollc.resume_rag.websockets;
 
 import java.io.IOException;
